Master

See merge request compsci308_2020spring/final_team11!19

+++ src/ooga/Model/Characters/Ghost.java
@@ -244,7 +244,7 @@ public class Ghost extends AbstractCharacter {
         return (int) (spriteImageView.getBoundsInParent().getMaxY() + spriteImageView.getBoundsInParent().getMinY())/2;
     }
 
-    public void setCenterY(int centerY) {
+    public void setCenterY(double centerY) {
         spriteImageView.setY(centerY);
         hurtBox.setY(spriteImageView.getBoundsInParent().getMinY());
